requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.0.0
accelerate>=0.20.0
flash-attn>=2.0.0
numpy>=1.21.0
pyyaml>=6.0
